+++ NEWS.md
@@ -1,5 +1,8 @@
 # criticalpath 0.2.1
 
+05/02/2022
+- Evaluate Redundancy function corrected for schedule without any relation.
+
 26/01/2022
 - Warn message from sch_add_relations_tibble function was corrected.